move VCS-specific code into own testing module
Allows easier testing of only a particular VCS type.
